fix: crash on Mac
On Mac, there is an issue where generating the window ifself as a webview causes the app clash.

Currently, we are avoiding this issue by creating bounds that cover the entire window.

However, since we must set `ClearColor` to `Color::NONE` and set `CompositeAlphaMode` to `CompositeAlphaMode::PostMultiplied`.,
we will revise this approach if an alternative solution is found.